pyproject.toml
[project]
name = "multi-agents-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.16",
    "aiosqlite>=0.21.0",
    "azure-identity>=1.21.0",
    "azure-keyvault>=4.2.0",
    "fastapi>=0.115.12",
    "geopy>=2.4.1",
    "greenlet>=3.2.0",
    "motor>=3.7.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.3",
    "python-dotenv>=1.1.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.1",
]

[dependency-groups]
test = [
    "httpx>=0.28.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
dev = []

[tool.bumpversion]
current_version = "0.1.0"
parse = "(?P<major>\\d+)\\.(?P<minor>\\d+)\\.(?P<patch>\\d+)"
serialize = ["{major}.{minor}.{patch}"]
search = "{current_version}"
replace = "{new_version}"
regex = false
ignore_missing_version = false
ignore_missing_files = false
tag = false
sign_tags = false
tag_name = "v{new_version}"
tag_message = "Bump version: {current_version} → {new_version}"
allow_dirty = false
commit = false
message = "Bump version: {current_version} → {new_version}"
moveable_tags = []
commit_args = ""
setup_hooks = []
pre_commit_hooks = []
post_commit_hooks = []
